Move applyCodePatch to patch.ts Improve agentEditFiles unit test
